# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Базовый набор FastAPI+SQLAlchemy+Postgres",
    "description": "Основа для приложения",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/dishes/": {
      "get": {
        "tags": [
          "Dishes"
        ],
        "summary": "Get Multi",
        "description": "GET `/dishes` - Получение списка блюд.",
        "operationId": "get_multi_api_v1_dishes__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cafe Id"
            }
          },
          {
            "name": "show_active",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Show Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishInfo"
                  },
                  "title": "Response Get Multi Api V1 Dishes  Get"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Dishes"
        ],
        "summary": "Create",
        "description": "POST `/dishes` - Создает новое блюда.",
        "operationId": "create_api_v1_dishes__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dishes/{dish_id}": {
      "get": {
        "tags": [
          "Dishes"
        ],
        "summary": "Get By Id",
        "description": "GET `/dishes/{dish_id}` - Получение информации о блюде по его ID.",
        "operationId": "get_by_id_api_v1_dishes__dish_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dish Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Dishes"
        ],
        "summary": "Update",
        "description": "PATCH `/dishes/{dish_id}` - обновление информации о блюде по его ID.",
        "operationId": "update_api_v1_dishes__dish_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dish Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CafeShortInfo": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Название кафе",
            "description": "Можно вводить в любом регистре"
          },
          "address": {
            "type": "string",
            "maxLength": 256,
            "title": "Адрес кафе"
          },
          "phone": {
            "type": "string",
            "maxLength": 16,
            "pattern": "^\\+?[0-9]{7,15}$",
            "title": "Телефон кафе"
          },
          "photo_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "address",
          "phone"
        ],
        "title": "CafeShortInfo",
        "description": "Схема с краткой информацией о кафе.\n\n* id: uuid.UUID (обязательное)\n* description: str (не обязательное)\n* name: str (обязательное)\n* address: str (обязательное)\n* phone: str (обязательное)\n* photo_id: uuid.UUID (не обязательное)"
      },
...
      "CustomError": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "CustomError",
        "description": "Схема кастомной ошибки для ответа API."
      },
      "DishCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "photo_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          },
          "price": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Price"
          },
          "cafes_id": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Cafes Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "price",
          "cafes_id"
        ],
        "title": "DishCreate",
        "description": "Схема для создание нового блюда.\n\n* `name` - string;\n* `description` - string;\n* `photo_id` - uuid;\n* `price` - int;\n* `cafes_id` - list uuid."
      },
      "DishInfo": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "photo_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          },
          "price": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Price"
          },
          "cafes": {
            "items": {
              "$ref": "#/components/schemas/CafeShortInfo"
            },
            "type": "array",
            "title": "Cafes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "price",
          "cafes"
        ],
        "title": "DishInfo",
        "description": "Схема с полной информацией о блюде.\n\n* `id` - int;\n* `name` - string;\n* `description` - string;\n* `photo_id` - uuid;\n* `price` - int;\n* `cafes` - list;\n* `is_active` - boolean;\n* `created_at` - date-time;\n* `updated_at` - date-time."
      },
      "DishUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "photo_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          },
          "price": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "cafes_id": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafes Id"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DishUpdate",
        "description": "Схема для обновления существующего блюда.\n\n* `name` - string;\n* `description` - string;\n* `photo_id` - uuid;\n* `price` - int;\n* `cafes_id` - list uuid.\n* `is_active` - boolean."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  },
...
}